fix: split ui and swagger handlers
