+++ package.json
@@ -9,7 +9,6 @@
     "mdns-resolver": "0.0.3",
     "reconnect-core": "1.3.0",
     "swagger-parser": "^6.0.5",
-    "swagger-ui-dist": "^3.22.1",
     "ut-swagger2-validator": "^0.3.1",
     "uuid": "3.3.2"
   },
@@ -21,6 +20,7 @@
     "nats": "^1.2.2",
     "rabbot": "^2.1.0",
     "sort-keys": "2.0.0",
+    "swagger-ui-dist": "^3.22.1",
     "tap": "^13.1.0",
     "ut-tools": "^6.11.0"
   },
